requirements.txt
argparse
matplotlib
numpy
Pillow
PyWavelets
scipy
torch
torchvision
tqdm
